Skin Shop Example
This example is simple skin shop, but it doesn't build due to a Serializable Sprite. In Unity Sprites DOES NOT serialize so I need to figure out diferent way of doing it instead of this way which was shown in a youtube tutorial. I am just keeping this as a short example of how to do the rest of the things. I.e. store data and use lists for multiple objects in project.